Separate database trigger into files
