+++ functions/src/index.ts
@@ -27,7 +27,17 @@ export * from './functions/exportHealthSummary.js'
 export * from './functions/getUsersInformation.js'
 export * from './functions/onHistoryWritten.js'
 export * from './functions/onSchedule.js'
-export * from './functions/onUserDocumentWritten.js'
+export * from './functions/onUserAllergyIntoleranceWritten.js'
+export * from './functions/onUserAppointmentWritten.js'
+export * from './functions/onUserBloodPressureObservationWritten.js'
+export * from './functions/onUserBodyWeightObservationWritten.js'
+export * from './functions/onUserCreatinineObservationWritten.js'
+export * from './functions/onUserDryWeightObservationWritten.js'
+export * from './functions/onUserEgfrObservationWritten.js'
+export * from './functions/onUserHeartRateObservationWritten.js'
+export * from './functions/onUserMedicationRequestWritten.js'
+export * from './functions/onUserPotassiumObservationWritten.js'
+export * from './functions/onUserQuestionnaireResponseWritten.js'
 export * from './functions/onUserWritten.js'
 export * from './functions/phoneNumber.js'
 export * from './functions/registerDevice.js'
