+++ functions/src/services/questionnaireResponse/dataUpdateQuestionnaireResponseService.test.ts
@@ -19,7 +19,7 @@ import {
   DrugReference,
 } from '@stanfordbdhg/engagehf-models'
 import { _defaultSeed } from '../../functions/defaultSeed.js'
-import { onUserQuestionnaireResponseWritten } from '../../functions/onUserDocumentWritten.js'
+import { onUserQuestionnaireResponseWritten } from '../../functions/onUserQuestionnaireResponseWritten.js'
 import { _updateStaticData } from '../../functions/updateStaticData.js'
 import { describeWithEmulators } from '../../tests/functions/testEnvironment.js'
