+++ functions/src/services/questionnaireResponse/kccqQuestionnaireResponseService.test.ts
@@ -9,7 +9,7 @@
 import { UserMessageType, UserType } from '@stanfordbdhg/engagehf-models'
 import { createKccqQuestionnaireResponse } from './createKccqQuestionnaireResponse.js'
 import { _defaultSeed } from '../../functions/defaultSeed.js'
-import { onUserQuestionnaireResponseWritten } from '../../functions/onUserDocumentWritten.js'
+import { onUserQuestionnaireResponseWritten } from '../../functions/onUserQuestionnaireResponseWritten.js'
 import { _updateStaticData } from '../../functions/updateStaticData.js'
 import { describeWithEmulators } from '../../tests/functions/testEnvironment.js'
