+++ functions/src/services/questionnaireResponse/registrationQuestionnaireResponseService.test.ts
@@ -16,7 +16,7 @@ import {
   QuantityUnit,
 } from '@stanfordbdhg/engagehf-models'
 import { _defaultSeed } from '../../functions/defaultSeed.js'
-import { onUserQuestionnaireResponseWritten } from '../../functions/onUserDocumentWritten.js'
+import { onUserQuestionnaireResponseWritten } from '../../functions/onUserQuestionnaireResponseWritten.js'
 import { _updateStaticData } from '../../functions/updateStaticData.js'
 import { describeWithEmulators } from '../../tests/functions/testEnvironment.js'
